Move WorkspaceList to user
